Sphereを計算グラフに変更
