+++ src/computationGraph/IVector3.ts
@@ -1,8 +1,10 @@
 import {IComputationNode} from './IComputationNode';
 import {IScalar} from './IScalar';
+import {Vector3Like} from './Functions';
 
 export interface IVector3 extends IComputationNode {
   readonly isVector3: true;
+  readonly vector3Value: Vector3Like;
   mul(other: IScalar | number): IVector3;
   dot(other: IVector3): IScalar;
   cross(other: IVector3): IVector3;
